+++ src/transpile/program_emit.rs
@@ -317,6 +317,7 @@ impl Transpiler {
             self.emit_runtime_helpers();
             self.emit_log_helpers();
             self.emit_fs_helpers();
+            self.emit_db_helpers();
         }
         Ok(())
     }
@@ -469,6 +470,7 @@ impl Transpiler {
             self.emit_runtime_helpers();
             self.emit_log_helpers();
             self.emit_fs_helpers();
+            self.emit_db_helpers();
         }
         self.indent -= 1;
         self.line("}");